% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-fit_models.R
\name{cleanAndReadCT2_out}
\alias{cleanAndReadCT2_out}
\title{FUNCTION_TITLE}
\usage{
cleanAndReadCT2_out(fName, colsExp)
}
\arguments{
\item{fName}{PARAM_DESCRIPTION}

\item{colsExp}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
