% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{createRpackage}
\alias{createRpackage}
\title{Rpackage Create}
\usage{
createRpackage(packname)
}
\arguments{
\item{packname}{Enter a direction, character, eg. 'selfpack'}
}
\description{
Use the funciton createRpackage to init env.
}
\details{
Use the funciton createRpackage to init env.
}
\examples{
packname='packsf' 
createRpackage(packname)
}
