% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patients.R
\docType{data}
\name{patients}
\alias{patients}
\title{World Health Organization TB data}
\format{
\subsection{\code{patients}}{

A data frame with 3 rows 2 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
patients
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
