% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayhello.R
\name{sayHello}
\alias{sayHello}
\title{Say Hello function}
\usage{
sayHello(name = "My Friend")
}
\arguments{
\item{name}{The name of the person you want to say hello to}
}
\description{
This is just a test function, it will way hello to you
}
\examples{
sayHello("Morpheus")
}
\keyword{test}
