% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage-package.R
\docType{package}
\name{testpackage-package}
\alias{testpackage}
\alias{testpackage-package}
\title{testpackage: Useful for greeting people}
\description{
This is a test package.
}
\author{
\strong{Maintainer}: Claus Bo Hansen \email{sbo@trolling.dk}

}
\keyword{internal}
